# swagger.json
{"info":{"title":"falcon API","version":"1.0.0","description":"Demonstrating how to use falcon api"},"host":"localhost:8080","basePath":"/","swagger":"2.0","paths":{"/api/v1/users":{"post":{"description":"Creates a new user","produces":["application/json"],"parameters":{"schema":{"$ref":"#/definitions/user"}},"responses":{"201":{"internalMessage":"New resource has been created"},"304":{"internalMessage":"Email already exists"},"500":{"userMessage":"Internal Server Error"}}},"get":{"description":"Returns all users","produces":["application/json"],"responses":{"200":{"description":"An array of users","schema":{"$ref":"#/definitions/user"}},"404":{"userMessage":"user not found"},"500":{"userMessage":"Internal Server Error"}}},"delete":{"description":"Deletes a single user","produces":["application/json"],"parameters":[{"name":"user_id","description":"user's id","in":"body","required":true,"type":"integer"}],"responses":{"204":{"description":"Successfully deleted"},"404":{"userMessage":"user not found"},"500":{"userMessage":"Internal Server Error"}}}},"/api/v1/clone":{"put":{"description":"Updates a single user","produces":"application/json","parameters":{"id":2,"in":"body","description":"Fields for the user resource","schema":{"type":"integer"}},"responses":{"201":{"description":"Successfully updated"},"500":{"userMessage":"Internal Server Error"}}}},"/api/v1/clone_available":{"get":{"description":"Returns the user to be cloned","produces":["application/json"],"responses":{"200":{"description":"returns a user","schema":{"$ref":"#/definitions/user"}}}}},"/*":{"get":{"description":"throws error 404","produces":["application/json"],"responses":{"200":{"description":"An array of users"}}},"put":{"description":"throws error 404","produces":["application/json"],"responses":{"200":{"description":"An array of users"}}},"post":{"description":"throws error 404","produces":["application/json"],"responses":{"200":{"description":"An array of users"}}},"delete":{"description":"throws error 404","produces":["application/json"],"responses":{"200":{"description":"An array of users"}}}}},"definitions":{"users":{"properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"country":{"type":"string"},"zip":{"type":"integer"},"phnum":{"type":"integer"},"alternateemail":{"type":"string"}}},"user":{"properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"country":{"type":"string"},"zipcode":{"type":"integer"},"phone":{"type":"integer"},"alteremail":{"type":"string"}}}},"responses":{},"parameters":{},"securityDefinitions":{},"tags":[]}